# Conflicts:
#	libs/nav/src/reactive/CAbstractPTGBasedReactive.cpp

+++ libs/graphslam/include/mrpt/graphslam/CGraphSlamEngine_impl.h
@@ -60,7 +60,7 @@ CGraphSlamEngine<GRAPH_t, NODE_REGISTRAR, EDGE_REGISTRAR, OPTIMIZER>::~CGraphSla
 	if (m_odometry_poses.size()) {
 		this->logStr(LVL_DEBUG, format("Releasing m_odometry_poses vector; size: %lu",
 					m_odometry_poses.size()));
-		for (int i = 0; i != m_odometry_poses.size(); ++i) {
+		for (size_t i = 0; i != m_odometry_poses.size(); ++i) {
 			delete m_odometry_poses[i];
 		}
 	}
@@ -2224,7 +2224,7 @@ void CGraphSlamEngine<GRAPH_t, NODE_REGISTRAR, EDGE_REGISTRAR, OPTIMIZER>::updat
 	// build the X, Y vectors for plotting - use log scale
 	std::vector<double> x(m_deformation_energy_vec.size(), 0);
 	std::vector<double> y(m_deformation_energy_vec.size(), 0);
-	for (int i = 0; i != x.size(); i++)  {
+	for (size_t i = 0; i != x.size(); i++)  {
 		x[i] = i;
 		y[i] = m_deformation_energy_vec[i]*1000;
 	}
